pyproject.toml
[tool.poetry]
name = "jointattention"
version = "0.1.0"
description = ""
authors = ["Igor Katselenbogen <underwoodkats@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.1.0"
numpy = "^1.26.2"
imageio = {extras = ["ffmpeg"], version = "^2.33.0"}
opencv-python = "^4.8.1.78"
pandas = "^2.1.3"
torchvision = "^0.16.1"
scikit-learn = "^1.3.2"
efficientnet-pytorch = "^0.7.1"
tqdm = "^4.66.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
